Moved xp to level updating from show_time
I personally didn't really experience performance issues when using the mod but as it stands the text for xp and gold to level was inside the show_time loop which may have caused performance issues if the UI attempted to rewrite it every second.

Now it will only attempt to update at the start of every round or when the player has spent 5 gold.